cleanup chain tests
